sakina-ai-service-main/requirements.txt
flask
flask-cors
transformers
torch
librosa
soundfile
tensorflow
scikit-learn
google-generativeai
python-dotenv
numpy
